requirements.txt
streamlit
spacy
pdfplumber
python-docx